subprocess: use a fifo instead of stdout for IO
Today, communication from a Kudu Java subprocess to the C++
SubprocessServer is done via the subprocess's stdout (the C++ message
protocol listens on the stdout FD of the subprocess). This is risky
because any amount of logging by the subprocess to stdout may end up
breaking the message protocol, preventing any calls to the subprocess
from succeeding.

This updates the Kudu subprocess module on both the Java and C++ side to
use a fifo for output instead of the subprocess's stdout. For Ranger,
this fifo will by default be placed in the Ranger config directory,
which, in some distributions (e.g. CDPD) will be unique per Kudu
process. Additionally, the name of this file is chosen to avoid
collisions with existing fifos.

This patch also updates the subprocess to share stdout and stderr with
the parent process by default, adding basic log4j configuration to the
Java subprocess to output to System.out. Here's a snippet of the output
from subprocess_proxy-test (which includes both Kudu's glog and stdout):

0 [pool-1-thread-1] DEBUG org.apache.kudu.subprocess.QueueUtil  - Message: org.apache.kudu.subprocess.InboundRequest@48c4ccb9 has been put on the queue
0 [pool-2-thread-1] DEBUG org.apache.kudu.subprocess.QueueUtil  - Message: org.apache.kudu.subprocess.InboundRequest@48c4ccb9 has been taken from the queue
1096 [pool-2-thread-1] DEBUG org.apache.kudu.subprocess.QueueUtil  - Message: org.apache.kudu.subprocess.OutboundResponse@6c275a0f has been put on the queue
1096 [pool-3-thread-1] DEBUG org.apache.kudu.subprocess.QueueUtil  - Message: org.apache.kudu.subprocess.OutboundResponse@6c275a0f has been taken from the queue
WARNING: Logging before InitGoogleLogging() is written to STDERR
I0327 23:35:20.583730 244412416 server.cc:279] Received an EOF from the subprocess
I0327 23:35:20.590333 244948992 server.cc:412] outbound queue shut down: Aborted:

This may not be the final form of our logging story for the subprocess
(e.g. maybe we'll want to log to a file, or maybe we'll want to try
folding it into the server's glog). But at the very least, this is
better than not having logs at all, which is the state of logging today.

I considered using Netty's FileDescriptor[1] to leverage the FDs from
calls to pipe(). I opted not to since that brought in some hairiness
around its usage of JNI (I saw an UnsatisfiedLinkError when I tried
writing to the FD).

[1] https://netty.io/4.0/api/io/netty/channel/unix/FileDescriptor.html

Change-Id: I63e91a090fe196713a013d28301c7980e452456c
Reviewed-on: http://gerrit.cloudera.org:8080/15574
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>