env: add a fifo class
This adds a basic Fifo class and a PosixFifo variant that encapsulates
the usage of mkfifo and associated functionality. My immediate need is
just to return the FDs of the opened fifo, so the interface is quite
bare. That said, this will be good enough to move the SubprocessServer
away from using the process stdout FD for IPC.

Change-Id: I23920457d695b84509937694daea53d61f445e27
Reviewed-on: http://gerrit.cloudera.org:8080/15584
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Kudu Jenkins