[build-support] introduce PARALLEL variable
This patch introduces PARALLEL variable into build-and-test.sh
Jenkins helper script.  The variable is used to explicitly set the
desired level of parallelism when building the project via make.
By default, the level of parallelism is determined by
`getconf _NPROCESSORS_ONLN`.

The motivation for this patch is the evidence of misreporting
on the actual number of available CPU cores by `geconf` if
running in some containerized environments like LXC.

The newly introduced variable named the same way as corresponding
variable in thirdparty/{build-definitions,build-thirdparty}.sh.

Change-Id: Ic1d5b36996d2b5e73f331a094c558abb97f895bc
Reviewed-on: http://gerrit.cloudera.org:8080/15591
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>