KUDU-3081 Add Kerberos support to MiniRanger
As integration tests using Sentry authz provider all depend on a
Kerberized mini cluster we also need to add Kerberos support to
MiniRanger to be able to parameterize the integration tests.

As adding Kerberos support proved to be trickier than expected this
patch contains several other improvements needed for this to work.

After configuring Kerberos for both MiniRanger and the Ranger subprocess
correctly I still ran into an error:

GSSException: No valid credentials provided (Mechanism level: Failed to find any Kerberos credentails)

This turned out to be due to a mismatch in forward and reverse lookup of
MiniRanger's hostname. To fix this I switched to using IP addresses in
the configuration instead of FQDNs.

This patch also changes the default "java" binary in Ranger client to be
the one in $JAVA_HOME/bin instead of the one in $PATH as dist-test has
mismatching versions and it wouldn't start otherwise. To use the one in
$PATH a user can still simply set it to "java", or provide a full path.

Change-Id: I32118780ad912791fe5e371004345428b6459549
Reviewed-on: http://gerrit.cloudera.org:8080/15601
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>