[thirdparty] root can run postgres
Sometimes Kudu development environment is run as a containerized VM
instance, and in such cases maintaining extra non-root users does not
make much difference.  Also, mini_postgres databases are not targeted
for production use anyway.  With that, it makes sense to allow to
run initdb and postgres process itself under UID 0 in the context of
mini_postgres as a part external mini-cluster test harness.

Change-Id: If6501648140bed3ba0df08a468cdf22a84f88cfc
Reviewed-on: http://gerrit.cloudera.org:8080/15613
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>