[ranger] Use unique loopback for MiniRanger
Other services use a unique loopback address on Linux systems to avoid
conflicting ports. This patch changes Ranger and Postgres to use this
unique loopback as well. Unfortunately as Ranger starts up Tomcat
without specifying the bind address, it always listens to all
interfaces.  Because of this the unique loopback trick is not enough in
itself and the randomized port number is also used in this case.

Change-Id: Id8833bb149fb2410f4f140b8fb67b1039a2f90d0
Reviewed-on: http://gerrit.cloudera.org:8080/15625
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>