Avoid calling Schema::find_column() once per RowBlock in columnar serialization
Prior to this patch, each row block being serialized in the columnar
format would result in a call to Schema::find_column(name) for each
projected column. That was relatively expensive, involving a hash
computation and string equality check, etc.

This changes the projection calculation to happen "up front" once per
Scan RPC and per-rowblock calls.

This optimization could also apply to the rowwise serialization, but I
found that the other overheads inherent in that code path are so high
that the find_column calls aren't particularly noticeable. Nonetheless
I left a TODO.

Change-Id: I1b683c7d6d6fe1026ee06c8b5ebfe2a5f1ee6cb1
Reviewed-on: http://gerrit.cloudera.org:8080/15678
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>