[subprocess] Fix shutdown behavior
Prior to this commit the subprocess didn't actually shut down when its
input stream was closed. This wasn't much of a problem on Linux where
PDEATHSIG is set on the child process, but as it's not supported on Mac,
whenever the master is shut down/crashes the subprocess stays alive.

This commit changes the behavior to exit when the reader future exits
(the loop is broken when EOF is received).

Change-Id: Ia84bff891eaff667d7b22e89c66f24980ad76d70
Reviewed-on: http://gerrit.cloudera.org:8080/15615
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>