cfile: change BlockBuilder API to yield a vector of Slices
When blocks are appended to cfiles at the IO layer, we already have the
ability to write multiple slices using a vectored IO. Prior to this
patch, the BlockBuilder API was restricted to returning a single slice,
whereas it would be more convenient in some cases to be able to return
multiple slices (eg separating the header from the data).

This new functionality is used by BinaryDictBlockBuilder to avoid an
extra copy in Finish().

Change-Id: Ifc7a5f148a4a43cedac2428f4c1a18d0f93a10db
Reviewed-on: http://gerrit.cloudera.org:8080/15042
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins