wire_protocol: change columnar serialization of varlen data to match Arrow
This changes the format of variable-length columns serialized on the
wire to match Apache Arrow instead of our internal column format. The
Arrow format consists of an array of n+1 offsets for n rows, such that
the data for cell 'n' spans offset offsets[n]...offsets[n+1].

The obvious advantage here is that clients can zero-copy into Arrow
structures since the format is compatible. The less obvious advantage is
that we are going from 16 bytes (sizeof(Slice)) to 4 bytes
(sizeof(uint32_t offset)) for each serialized string, so this should be
a savings even for non-arrow users of the API.

This patch also adds some more sanity checking of the wire format in the
client API so that it's not the responsibility of the caller to guard
against malicious servers.

Change-Id: Iadf728744feb83f5980e62bea4fd7634a1a52467
Reviewed-on: http://gerrit.cloudera.org:8080/15660
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>