columnar_serialization: avoid preallocating 8MB per column
Previously we would reserve an 8MB buffer for every column of data to be
scanned. This wouldn't scale well for high number of concurrent queries
with lots of columns.

The new approach is to use the configured batch size and apportion that
memory budget across the columns based on the size of those columns.
It's not 100% accurate but at least shouldn't overshoot by hundreds of
MB like the prior approach.

Change-Id: I9b7ff78547792acbd975a606a02ec388dba3a8e8
Reviewed-on: http://gerrit.cloudera.org:8080/15679
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins