columnar_serialization: fix optimized GCC build
It seems that GCC with -O1 or higher is less lenient about the pointer
types passed as the first argument to the _mm_gather* intrinsics. This
was causing the Ubuntu 18 release build to fail.

I was able to reproduce the failure and fix on my own Ubuntu 18 box
using a fastdebug build.

Change-Id: I6b69470b238dae7a33fdd6b44cb8be57a26501d7
Reviewed-on: http://gerrit.cloudera.org:8080/15680
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>