net: make Sockaddr more generic for other address families
This paves the way for supporting other socket address types: namely
IPv6 and Unix domain sockets. This patch doesn't add any new support,
but increases the size of Sockaddr appropriately and making the methods
more generic.

The approach here is to use a union inside of the Sockaddr class which
could store an address of any type. An alternative (taken by a long-ago
attempt at unix sockets) would be to make Sockaddr an abstract base
class with derived classes for each socket type. I went with this
approach since all of the POSIX socket APIs expect this "tagged union"
approach -- the same APIs are used whether we want to talk to a unix
socket or an IPv4. The downside of this approach is a slightly larger
Sockaddr object footprint, but we shouldn't have a lot of these on the
heap.

One functional change: the default constructor for Sockaddr now creates
an "uninitialized" instance, whereas it used to create a wildcard
address. That's now an explicit static method.

Change-Id: Ifff57e13a0e9b86b191d776ead67371b1e2ed4e3
Reviewed-on: http://gerrit.cloudera.org:8080/15690
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>