[catalog_manager] reduce contention in ScopedLeaderSharedLock
While troubleshooting one performance issue if running a big cluster
with large number of tables and high rate of ConnectToMaster requests,
in the logs I noticed many reports like the following:

  0323 03:59:31.091198 (+607857us) spinlock_profiling.cc:243]
  Waited 492 ms on lock 0x4cb0960. stack:
    0000000002398852
    0000000000ad8c69
    0000000000aa62ba
    000000000221aaa8
    ...

which translates into
    (anonymous namespace)::SubmitSpinLockProfileData()
    master::CatalogManager::ScopedLeaderSharedLock::ScopedLeaderSharedLock()
    master::MasterServiceImpl::ConnectToMaster()
    rpc::GeneratedServiceIf::Handle()
    ...

From the code it became apparent that the lock in question was
  std::lock_guard<simple_spinlock> l(catalog_->state_lock_);
in ScopedLeaderSharedLock() constructor.

As far as I can see, there is no need to access master's Raft consensus
information (which itself might wait on its internal locks if there is
corresponding Raft-consensus activity) under the catalog's state lock.

This patch shortens the critical section with catalog's state lock held
when constructing CatalogManager::ScopedLeaderSharedLock instance.

Change-Id: I3b2e6866a8a0d5bda9e2b1f01e0668427de60868
Reviewed-on: http://gerrit.cloudera.org:8080/15698
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins