[test] fix TestCorruptKerberosCC scenario
On macOS, the SecurityITest.TestCorruptKerberosCC scenario was failing
with errors like below:

  src/kudu/integration-tests/security-itest.cc:430: Failure
  Value of: s.ok()
    Actual: true
  Expected: false
  Google Test trace:
  src/kudu/integration-tests/security-itest.cc:419: Truncating ccache at
    'security-itest.0.SecurityITest.TestCorruptKerberosCC/krb5kdc/krb5cc' to 500

It seems Kerberos cache truncated to 500 bytes is big enough to
successfully extract necessary data on macOS.

This patch decreases the size of the corrupted client cache to 266
bytes, so now the scenario passes on macOS as well.  In addition,
the truncation size is now selected randomly.

Change-Id: I765e9f1f6cd208f86bc321e962588982f9e01447
Reviewed-on: http://gerrit.cloudera.org:8080/15786
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>