tablet: set up RowSetKeyProbe fields in initialization list
For some reason, setting these fields in the body of the constructor
generated a lot more code than setting them in the constructor list.
Namely, the setting of a unique_ptr<> generated a check whether the
pointer was already set, and code to deallocate the old version if it
was.

I'm not sure why clang couldn't figure out that it couldn't possibly be
set -- perhaps something to do with inheritance (eg maybe a subclass
would have set the field in its constructor before chaining upwards?).
Either way, this was taking 2-3% CPU and dropped after this change.

Change-Id: Ibe46e00f5ca8c13c7366df4bee04e7c6e840b3a0
Reviewed-on: http://gerrit.cloudera.org:8080/15790
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Tested-by: Kudu Jenkins