transaction: avoid locking for TransactionState::timestamp_
This field is already synchronized externally for mutations by the transaction
lifecycle (we never have a case where multiple threads are trying to
concurrently assign a timestamp to a transaction). The timestamp might
be concurrently read (eg by ToString() for the /transactions web page)
but an atomic variable is sufficient for this use case.

I saw a couple percent of the Apply thread's CPU going to managing this
locking, which should go away with this change.

Change-Id: I2c6bc8b85a066426bcf4bc2d1c058bc75c1c38ec
Reviewed-on: http://gerrit.cloudera.org:8080/15791
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins