clients: prefer tservers in the same location when multiple are local
be68ce8 introduced the behavior that mini-servers that are local to a
C++ client are treated as local, even on Linux. As such, a
location-aware mini-cluster would not honor client location awareness,
since all mini-servers would be considered local.

While this surfaced as flakiness of location_awareness-itest, this is an
actual product deficiency, albeit one we don't expect to be very common:
if multiple servers are colocated with a client, and the servers are
assigned different locations, those locations should be honored when
picking tablet servers.

Without this patch, location_assignment-itest is ~24% flaky. With it, it
passed in debug mode 300/300 times.

The same logic is updated in the Java client.

Change-Id: I6a0dbac073ba8a9f80a3f175157a09d06037b0df
Reviewed-on: http://gerrit.cloudera.org:8080/15794
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>