[tools] multiple tablet ids in 'local_replica delete'
This patch updates the 'local_replica delete' tool to allow for
multiple tablet identifiers to be specified and processed at once.
The rationale behind this change is that opening tablet server's
metadata takes significant time, and it's possible to reduce
the overall latency if removing multiple tablet replicas at once
after the metadata has already been opened.

A new test to verify the new functionality has been added as well.

Change-Id: If0e509d1775be2a728e4e3b10c724c1f15a96ec1
Reviewed-on: http://gerrit.cloudera.org:8080/15909
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>