KUDU-2817: Upgrade glog to 0.4.0
We are currently on glog 0.3.5 which is from 2017. We also have 4 patches
on top of that version in thirdparty.

This patch upgrades us to glog 0.4.0 which was released in 2019 and
removes 3/4 patches that are now included in the release.

I needed to adjust the remaining patch to apply and fix some usage of
the CHECK macros to wrap in parenthesis to prevent macro expansion
issues as well.

I also adjusted the build scripts to use cmake instead of autoconf so
that I could disable building the tests.

Change-Id: I652a852d040fcd3d974002382735610c8621c413
Reviewed-on: http://gerrit.cloudera.org:8080/15924
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>