KUDU-2817: Upgrade snappy to 1.1.8
We are currently on snappy 1.1.4 which is from 2017 and the latest
release is from January 2020.

This patch upgrades to snappy 1.1.8 which pulls in various fixes and
improvements. See the release notes for all the changes:
https://github.com/google/snappy/releases

I adjusted the build scripts now that snappy uses cmake instead of
autoconf. I used the parameters passed to lz4 for inspiration.

Change-Id: I19e4eda122c0ea957bbe504fc7de681bb72068cd
Reviewed-on: http://gerrit.cloudera.org:8080/15927
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>