[client] Remove allocator param for direct Bloom filter predicate
Direct Bloom filter predicate in kudu client API is meant for consumption by
Impala.

Supplying the allocator that's associated with Bloom filter is turning out to
be tricky for Impala. For instance, it doesn't want to support Clone() and
concerned with the allocator being used by kudu client for
allocation/deallocation of memory.

Allocator associated with the Bloom filter is only used for cloning a predicate
and the support for cloning is only used for internal testing purposes.

So this change removes the allocator parameter from the direct Bloom filter
predicate API. For cloning the direct Bloom filter, it uses the default
allocator instead.

Since this API is marked as advanced/unstable, no compatibility concerns.

Change-Id: I7b57a6cacc8a5b6243787df8f896b74aa04f3916
Reviewed-on: http://gerrit.cloudera.org:8080/15947
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>