[catalog] remove CatalogManager::CheckOnline()
The implementation of CatalogManager::CheckOnline() was there to verify
that the catalog is running.  As it turns out, callers of the methods
where CheckOnline() was invoked always (except for the code in
master_path_handlers.cc) create ScopedLeaderSharedLock instance prior
to a call with a follow-up verification of the catalog's status (e.g.,
see MasterServiceImpl).  In other words, CheckOnline() was always
performing a verification that has already been done.

This patch removes CatalogManager::CheckOnline() along with the extra
verification of the catalog's status.

Change-Id: I00dd5d1f849d081e6a36190871e352e98f16f0a2
Reviewed-on: http://gerrit.cloudera.org:8080/15977
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>