[scanner] output the essential error message first
While digging through logs as they seen in kudu-tserver Web UI,
I found it hard to get to the essence of the warning message on
setting up a new scanner: it contained a very long string prefix
with the information about the scan projection.

This patch addresses that, moving the essential information on the
status into the beginning of the message.

Otherwise, this patch doesn't contain any functional modifications.

Change-Id: I06e4ed516dcd2b520f26ddc580aed00be239df04
Reviewed-on: http://gerrit.cloudera.org:8080/15999
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>