KUDU-3148: [test] Add Java client metrics
This patch adds the basic tools for instrumenting the Java
Kudu client with metrics. The primary purpose for this patch
is greater visibility and validation in tests. However, it leverages
micrometer for the metrics to allow for expanded usage and use
cases in the future (e.g JMX publishing, periodic diagnostic
logging, etc).

In this patch I add some counter metrics for RPC requests,
retries, and responses. I leverage these metrics in
TestScanToken to validate the current behavior of ScanToken
generation and use. A follow on patch will use these metrics
more to validate improvements to the ScanToken.

I did not use or enhance the existing Statistics.java
implementation because it didn’t appear to be flexible and
powerful enough for general purpose metrics without a lot of
work (effectively writing something like micrometer). Outside
of the additional work, Statistics.java is public API which makes
it more difficult to change.

Change-Id: I5c63835dd717c2c1e1dca06ed5dea3c2cadcd018
Reviewed-on: http://gerrit.cloudera.org:8080/16067
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins